since we need just some parsers, use this older ffmpeg variant to reduce size
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@33467 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/minidlna/ffmpeg-3.1/libavcodec/vp6dsp.c
@@ -1,4 +1,7 @@
-/*
+/**
+ * @file
+ * VP6 DSP-oriented functions
+ *
  * Copyright (C) 2006  Aurelien Jacobs <aurel@gnuage.org>
  *
  * This file is part of FFmpeg.
@@ -18,16 +21,11 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-/**
- * @file
- * VP6 DSP-oriented functions
- */
-
 #include "libavutil/common.h"
 #include "vp56dsp.h"
 
 
-void ff_vp6_filter_diag4_c(uint8_t *dst, uint8_t *src, ptrdiff_t stride,
+void ff_vp6_filter_diag4_c(uint8_t *dst, uint8_t *src, int stride,
                            const int16_t *h_weights, const int16_t *v_weights)
 {
     int x, y;
